View.wrapContent method defined. View rendering logic explained.
The "wrapContent" method wraps some HTML string (typically that
returned by the "generate" function) into an HTML serialization
of the view's DOM element. Its main use is for the implementation
of the "generate" method of container views that need to retrieve
the HTML content produced by their children. They may now do it
through something like:

generate: function(cb) {
 var self = this;
 this.innerView.generate(function(err, innerHTML) {
  var outerHTML = self.innerView.wrapContent(innerHTML);
  return cb(err, self.template({ child: outerHTML }));
}

They must also call View.setElement in their setContent function
to update the DOM element of their children and re-bind potential
event handlers to the new element:

setContent: function(html) {
 UIItem.prototype.setContent.call(this, html);
 this.innerView.setElement(this.el.firstChild, true);
}

This guarantees that the inner view produces the same result as
if it was used outside of the container view.

This logic is clarified in the code.