DynamicContainer: sets the actual view based on the model to render.
The detailed view associated with a list of items typicall depends
on the type of items to render. The most common scenario is that
the list only contains items of a certain type and thus the detailed
view is known a priori.

That may not always be the case though, as datasources may return
mixed content. In that case, the detailed view typically depends
on the type of item to render. The DynamicContainer view makes it
possible to implement that logic, reporting the creation of the
view used to render the underlying model to the time when the view
needs to be rendered and using a factory to create the view based
on the model to render.

See code for usage example.