Scroller added to phone's toolbar once element has the right size.
The scroller was only "more or less" added once the toolbar had been
resized, leading to unexpected behavior from time to time. The bug
was that the scroller was added after a 10ms timeout. That was
usually enough for the toolbar to be resized in the meantime but
that was not a guarantee.

The toolbar now resizes its content first and inserts the scroller
once that is done.