Adapted mediaplayerlib/FactoryMedia to preserve video aspect ratio
More precisely, mobile browsers may not respect the intrinsic ratio
of a video element when they render a <video> tag as opposed to what
the HTML5 spec suggests. This means that setting the <video> tag to
arbitrary width and height may distort the video when it plays. That
typicaly happens today with the default Web browser of Android Ice
Cream Sandwich.

The mediaplayerlib may now automatically adapt the dimensions of the
media element it returns based on the intrinsic VideoObject
dimensions, if so requested through the "adjustSize" option. It
also returns the dimensions of the element it creates along with
the HTML markup of the element.

The FactoryMedia view was adjusted to expose the actual size of the
created element to the HTML template associated with the view.

Check the comments at the beginning of the FactoryMedia view for
more details about dimension options and how they can be used.