Phone adapted: iScroller may now be linked to an element within a view.
Phone views now accept a "scrollerSelector" option. When present,
the scroller gets added to the selected element.

NB: other adapters do not (yet) provide that feature. It is mandatory
in the case of mobile phones for views that contain a <video> tag,
because the <video> tag fails to detect clicks when it's contained
within a section for which iScroll is activated in the default Web
browser of Android ICS.