"shown" and "hidden" events triggered. Scroller refreshed on "shown".
The "shown" and "hidden" events are triggered after the "show" or
"hide" functions have been invoked. Container views propagate the
events down to their children to ensure they receive it.

This behavior gives a hook to views to update their layout when
the view moves from a hidden state to a displayed state. Typically,
this is needed for views that embed iScroll to provide a custom
scroller. iScroll listens to various events and refreshes the
dimensions of the scroller accrodingly. When these events occur
while the view is hidden, the scroller simply disappears. On top
of this, the dimensions of the viewport might have changed when a
view is displayed again (e.g. following an orientation change),
so the scroller needs to be refreshed in any case.

The phone UIElement adapter takes care of all this. Containers must
either derivate from UILayout or ensures that they propagate
"shown" and "hidden" events to their children.