New ImagesLoader view for multiple images. New "load" event triggered.
The ImagesLoader view is similar to the ImageLoader view except it
takes an array of images as input. The images extractor may be
specified in the getImages parameter. See inline comments for details.
By default, the view extracts all <img> tags.

Each image may be updated before the "loading" class gets applied if
the processImageEl function is passed as parameter. This function is
typically useful to add a spinner while the image is loaded in the
background.

The commit also includes the foundations for a "load" event triggered
by all views whenever the view is done loading its assets. The event
is triggered at the end of the "enhance" function by default but that
behavior may be overridden in derivated classes by setting the
"customLoadEvent" flag. Containers typically trigger the "load" event
when all of its children views have triggered their. Images loader
views only trigger the "load" event when images are done loading.

This event can be used to display a spinner up in the hierarchy of
the application, e.g. in the title bar or as an overlay in the
middle of the screen. The "load" event must be fire once and only
once each time a view is rendered.

Note from the author: the "load" event has not been thoroughly tested
yet. I am about to use it in the Sleek template for the pictures
listing view.