Loading class automatically handled by views when requested.
The "loadingClass" view initialization option may now be passed
around to all views to tell the view to add the class to the view's
element between the time when the beginning of the "enhance" function
and the "load" event, in other words while the contents of the view
are loaded by the browser.

This is typically only useful when images are loaded in the background.

The "dataLoadingClass" initialization option may also be passed to
list views to tell the view to add the corresponding class to the
view's element while data is sync-ed. It is "loading" by default.

The data loading classname and the view loading classname should be
different to avoid conflicts (when data is sync-ed, the class would
be removed from the view's element, but render might have been called
in the meantime, meaning the view loading classname might have been
added).