Possibility to disable scroller during startup with "disabled" flag
iScroll exposes a "disable" function that now gets called when the
UIElement is created with a "disabled" flag in the scrollOptions
property.

This forces the creation of a disabled scroller, which may still
be useful when the scroller is used to trigger explicit transitions
between pages (typically what happens a toolbar that has continuation
points). In turn, this makes it impossible to slide the toolbar
(which may or may not be viewed as a good thing) but more importantly
it prevents running into touchstart/click event issues on Android
that make "click" events fire twice.