ImagesLoader: Fixed missing call to base enhance if no images to load
When there are no images to load in the background (which may happen
when the view is used to wrap an article whose content we don't know
anything about a priori), the view failed to call the base class
"enhance" function. In turn, that means the UIElement.enhance function
was not called which, in the case of phones, meant the scroller did
not appear.