ImagesLoader: Fixed missing call to "enhance" and images extraction in Zepto
When there are no images to load in the background (which may happen
when the view is used to wrap an article whose content we don't know
anything about a priori), the view failed to call the base class
"enhance" function. In turn, that means the UIElement.enhance function
was not called which, in the case of phones, meant the scroller did
not appear.

Latest version of Zepto introduced by a previous commit also has some
interesting features: $.map(obj, cb) is different from $(obj, cb).
The order of parameters passed to the callback function is not the same.
Images extraction failed as a consequence. I switched to the $.map
form so that code is compatible with jQuery as well.