View containers now always bubble "load" events from children views
The "load" event was only triggered once when the container view
was first rendered. Children view may be rendered again later on
though. Container views that derivate from Layout now bubble the
"load" events from their children when they occur oustide of the
context of a rendering of the container view. The events are bubbled
individually since the container view has no way to know when it
receives a "load" event from one of its children whether there will
be another one or not.

Bubbling the "load" event makes it easy to implement a callback
function that gets called whenever the main view of an application
has been fully refreshed. Typically, this is useful for applications
that start by rendering the main view without data and then refresh
portions of the main view once data is available.