Framework path resolved against main script; config read by optimizer
The Framework did not support declarations such as:
<script type="text/javascript"
  src="js/lib/framework/adapters/ios/bootstrap.js"
  data-main="js/main"></script>

... because the "js/main" value makes require.js use "js" as base
folder for all dependencies resolution, while the joshlib plugin in
the Joshfire framework thought the base folder was the one of the
HTML file.

The Framework is smarter and resolves the framework folder against the
base folder set by the main script, provided the path that leads to
the main script does not contain any "../" subpath.

Similarly, while optimizing, the optimizer did not check whether the
main script file defines some "require.js" configuration or not.
This is actually useful to implement a lightweight adapter system in
the app. The optimizer now uses require.js "mainConfigFile" option
to read the config if defined. This makes it possible to have code
such as:
 requirejs.config({
  paths: {
    'view/class': 'view/class.tablet',
    'base': '.'
  }
 });
... in one of the main entry points of the application to define
specific classes for that entry point, and to have the optimizer not
miss the adapted class.

I re-generated the bootstrap files of the different adapters
accordingly and took the opportunity to update code in these files
to follow our guidelines.