Changes the mechanics of the Layout
... Although it should be backwards-compatible.
Basically, layouts now allow the user to define
the emplacement in the DOM where the child element
will be inserted. This is done by accessing the
"children" property which has been added to the
layouts' rendering global scope.
Cardpanel was correct, though. Some of the previously
used logic which was in Layout has been transferred
to CardPanel instead, which feels more appropriate.