List view and base collection: "load more" improved and Woodman logs
The base collection wrapper now triggers "load:more" and "load:last"
events and maintain a "hasMore" flag initially set and reset when
the collection receives fewer items than the number of items requested
in a fetch request.

I re-wrote the description of the List view to detail the options
available and simplified the "load more" feature. The code may still
need to be updated for other scrollers as the "onScroll" event is
kind of bound to iScroll way of thinking.

The "load more" feature does not create any additional markup anymore,
but flags the root of the view with a "loading" class name (which may
be specified when the view is created). This should allow all or
next-to-all possible animations to happen within the view to warn the
user that data is being loaded.

First few logs added with Woodman to track things down.