fetchMore: new function in Collection, fixed in TV Grid/Verticallist
The base collection class now exposes a "fetchMore" function that
triggers a new fetch with the appropriate skip/limit parameters.
That function is used by the List view to retrieve more items.
The function does not take any callback, the caller should listen to
the "load" events it is interested in.

Previous commits broke the TV versions of Grid and VerticalList that
expected List to expose "loadMore" in another way. This commit fixes
the problem but note scrolling may still break when these views are
used:
- user may not be able to scroll to the last added images in a grid
- active item may go berserk in a vertical list when new items are
added (pressing up or down again fixes the problem).

As far as I can tell, these bugs existed before I made changes to the
List view.