TV Grid/VerticalList: fixed nav after "load more". Woodman added
In both classes, problems arised when new items were added to the
collection because the "resize" was computed too early.

The List view only renders children on the tick that follows the
insertion of the children in the list to ensure all new children get
rendered at once. There was no proper way to detect the end of the
rendering. The List view now triggers a "render:children" event once
the rendering of the new children is over. The Grid and VerticalList
views now listen to that event and update themselves accordingly.

Woodman traces added to ease debugging and maintenance.