Samsung TV adapter views dropped when possible, simplified when not
Most of the views adapted for Samsung TV were not really different
from the TV adapter versions of these views in the end. Most of the
code ended up in the adapter for historical reasons (different things
were tried out before rolling back to a more regular behavior).

In the interest of easing maintenance, I dropped the views that did
not bring anything and simplified those that could be simplified.
The only really view that remains highly specific to Samsung is the
VideoPlayer view.