Added "initializeLogId" in base View class. Dropped "utils/uid".
The "initializeLogId" function sets "this.logid" that can be used
to trace events that occur on an instance of a view with Woodman.
The "logid" uses Backbone's view's "cid" value as basis but can be
completed with a more meaningful name when the view is created to
be able to quickly tell what the view is for in the logs.

The function is called by all views that use Woodman before the first
call to "logger.log".

I dropped the "utils/uid" that I had introduced recently since
Underscore already exposes a "uniqueId" method that basically does
the same thing.