Consistent "update" mechanism when model/collection get changed
The "setModel" and "setCollection" methods used to take a second
parameter that forced a complete rendering of the view. That
parameter has now been replaced by a parameter that triggers the
"update" mechanism of the view when the model and/or collection is
modified.

By default, the "update" mechanism forces a complete rendering of
the view but only if needed. For instance:
- for a list, this is when the collection gets changed but not when
individual items in the collection trigger a change, since these
changes get caught by the views that are bound to the element.
- For a dynamic container, changing the number of items in the
collection does not necessarily mean that the "view factory" has to
be called again.