TV adapters: isFocusedElement function and Woodman traces added
Within the TV world, the views maintain the view that currently has
the focus in a static variable attached to the TV version of the
UIElement class. The "isFocusedElement" function returns true when
the current view has the focus.

The whole focus/blur mechanism needs to be revisited later on. As
things stand, it duplicates what the browser does and makes things
uselessly complicated. We should rather rely on HTML5 focus features
whenever possible.