Further simplified Samsung TV adapters. Bootstrap files updated.
I dropped the old adapted View class since it did not contain any
specific code. It can always be added back later on if needed.

The UI Element class borrowed a lot from the TV version. I reduced
it to a bare minimum that only forces the "translate3d" option to
"false".