Base View class: view now maintains a "hidden" flag and option
The "hidden" flag is set when the view is hidden through a call to
"hide". Subsequent calls to "hide" do nothing (well, they used to
do nothing as well before, but they were still calling jQuery's
hide function which resulted in DOM updates that are best avoided
to reduce the likelihood of flickering).

The "hidden" option may be given when the view is created to alert
that the view is to be created in a "hidden" state. The view's root
element receives a "display:none" style attribute in that case when
the view is rendered.

Similarly, the "show" function only does something when the view is
in a hidden state.

Note that this means that the "hidden" state now needs to be handled
by the code logic, which is good practice in any case. In particular,
the HTML template of the view must not contain a "display:none" style
attribute as the view would think it is in a non hidden state and
thus the call to "show" would simply not perform any action.