FactoryMedia: Woodman traces, warning on stupid generation code
The FactoryMedia view now reports computed width and height of the
requested video to Woodman.

The view includes code in "generate" that applies the requested
percentage to the width/height of the view when these dimensions
have not been provided when the view was created. That is rather
stupid since, by definition, the view has not yet been rendered
when "generate" is called so its width is always 0. I left the
code in with a warning trace in case I missed something.