Added devicedetect lib...
The point is to let the developper define multiple implementations of a class for a panel of devices. The plugin lets the dev require a file and will load the correct ones according to tests designed by the dev. These tests/mappings will be covered in another commit (namely on the boilerplate project.)

devicedetect provides a runtime object to the dev for him to run tests. It does so by parsing the client's UA using UAparser (hence its inclusion) when it is run unoptimized. Otherwise, UAParser is still used in nodejs to - if necessary - parse a UA string that has been provided as a CLI argument to the optimizer script.

This is necessary since we can't fully rely on the adapter variable anymore - we seek to handle multiple formfactors, os or constructor which gets a tad complicated.