Added commander support. This allows for CLI options...
Which enables us to specify options in whichever order we want and be a little more explicit about it. Retro compatibility tests remain to be done for this version.

Also added support for several new options which allow the user to specify a device, formfactor, etc. without having to specify a complete user agent.