Dynamic container: restricted updates to "final" collection events
The dynamic container may decide to update its child based on changes
made to an underlying collection. The class used to listen to all
possible change events, including e.g. "add", but these events may
be followed by other similar events.

If the class decided to create a List once it received a first "add"
event, the List would be initialized with a complete collection
(all items would already be present) but would still receive the
following "add" events, making it think that new items are being added
to the collection.