Card panel: ensured only "first" child is shown at first
The view rendered all children without selecting the current child,
meaning that all cards typically appeared on screen until a call to
"showChild" eventually hid away other cards.

The view now picks up the "first" child, or rather the first name
it finds when it enumerates the keys of the list of children. Note
that this order is not guaranteed in theory, it just happens to be
fairly consistent across runtimes...