Transition/Swipe: unknown transitions, event handlers undelegation
The transition panel crashed when a non supported transition was
requested. Unknown transitions are now properly converted to "no"
transition.

The touch event handlers that the slidepanel view attached to the
slide container are now properly unbound when the content is replaced
one way or the other to enable garbage collection and avoid side
effects.

The transition panel also triggers a "transitionend" event (was
"transitionEnd") to be more consistent with the DOM event it relates
to.